+++ DBServer/DBServer/DBServer.csproj
@@ -95,9 +95,7 @@
     <Compile Include="DBController.cs" />
     <Compile Include="DBServerConfig.cs" />
     <Compile Include="DBServerMain.cs" />
-    <Compile Include="ErrorCode.cs" />
     <Compile Include="MongoDBManager.cs" />
-    <Compile Include="PacketData.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RedisManager.cs" />
@@ -106,6 +104,12 @@
   <ItemGroup>
     <None Include="App.config" />
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\Library\CommonLibrary\CommonLibrary.csproj">
+      <Project>{64086658-8a99-4dcc-a083-b00e6fbb46d4}</Project>
+      <Name>CommonLibrary</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
