+++ Library/CommonLibrary/CommonLibrary.csproj
@@ -4,4 +4,8 @@
     <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="MessagePack" Version="1.7.2" />
+  </ItemGroup>
+
 </Project>
