+++ Library/CommonLibrary/TcpPacket.cs
@@ -1,10 +1,48 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using MessagePack;
+using System;
 
-namespace CommonLibrary
+namespace CommonLibrary.TcpPacket
 {
-    class TcpPacket
-    {
-    }
+	[MessagePackObject]
+	public class PacketHeader
+	{
+		[Key(0)]
+		public int PacketId;
+		[Key(1)]
+		public int BodySize;
+
+		public PacketHeader()
+		{
+			PacketId = 0;
+			BodySize = 0;
+		}
+
+		public PacketHeader(int bodySize, int packetId)
+		{
+			BodySize = bodySize;
+			PacketId = packetId;
+		}
+	}
+
+	[MessagePackObject]
+	public class LoginReq
+	{
+		[Key(0)]
+		public string UserId;
+		[Key(1)]
+		public Int64 Token;
+	}
+
+	[MessagePackObject]
+	public class LoginRes
+	{
+		[Key(0)]
+		public int Result;
+	}
+
+	public enum PacketId
+	{
+		LoginReq = 101,
+		LoginRes = 102
+	}
 }
