+++ Library/LogicLibrary/LogicLibrary.csproj
@@ -10,6 +10,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\CommonLibrary\CommonLibrary.csproj" />
     <ProjectReference Include="..\NetworkLibrary\NetworkLibrary\NetworkLibrary.csproj" />
   </ItemGroup>
