+++ Library/LogicLibrary/UserManager.cs
@@ -1,11 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using NetworkLibrary;
+﻿using NetworkLibrary;
 
 namespace LogicLibrary
 {
-    public class UserManager : IUserManager
+	public class UserManager : IUserManager
     {
 		bool IUserManager.IsUserExist(ClientSession owner)
 		{
