+++ Library/NetworkLibrary/NetworkLibrary/BufferManager.cs
@@ -1,7 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Net.Sockets;
-using System.Text;
 
 namespace NetworkLibrary
 {
