+++ Library/NetworkLibrary/NetworkLibrary/BytePacker.cs
@@ -1,7 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using CommonLibrary.TcpPacket;
 using MessagePack;
+using System;
 
 namespace NetworkLibrary
 {
