+++ Library/NetworkLibrary/NetworkLibrary/Defines.cs
@@ -1,12 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Runtime.InteropServices;
-using MessagePack;
-
+﻿
+// 네트워크 라이브러리 내부에서 사용할 상수 값들 정리.
 namespace NetworkLibrary
 {
-    internal class Defines
+	internal class Defines
     {
 		public static readonly int HeaderSize = 3;
     }
