+++ Library/NetworkLibrary/NetworkLibrary/NetworkLibrary.csproj
@@ -11,4 +11,8 @@
     <PackageReference Include="NLog.Extensions.Logging" Version="1.0.0-rtm-rc2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\CommonLibrary\CommonLibrary.csproj" />
+  </ItemGroup>
+
 </Project>
