+++ Library/NetworkLibrary/NetworkLibrary/SocketAsyncEventArgsPool.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Net.Sockets;
-using System.Text;
 
 namespace NetworkLibrary
 {
